requirements.txt
asgiref==3.2.3
astroid==2.3.3
autopep8==1.4.4
certifi==2019.11.28
chardet==3.0.4
Django==3.0.2
django-debug-toolbar==2.1
django-shortuuidfield==0.1.3
djangorestframework==3.11.0
et-xmlfile==1.0.1
idna==2.8
isort==4.3.21
jdcal==1.4.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysqlclient==1.4.4
openpyxl==2.6.3
Pillow==6.2.1
pycodestyle==2.5.0
pylint==2.4.4
PyMySQL==0.9.3
python-memcached==1.59
pytz==2019.3
qiniu==7.2.6
requests==2.22.0
shortuuid==0.5.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.7
wrapt==1.11.2
